Custom tapping term for fingers
Add ability to increment/decrement tapping term differences for index
and ring/pinky fingers. Add two unsigned integers for the difference to
apply to the fingers, add custom keys to increment/decrement these
values and add custom key to print both these values and the dynamic
tapping term difference. Also add convenience methods to convert a
uint16 to a string and sending the string only when the feature is
enabled.